backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI & Image Processing
openai==1.3.6
Pillow==10.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0

# HTTP & Validation
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1